+++ app/src/test/java/com/sun/weatherapp/testutil/TestDummies.kt
@@ -1,0 +1,84 @@
+package com.sun.weatherapp.testutil
+
+import com.sun.weatherapp.data.model.CurrentWeather
+import com.sun.weatherapp.data.model.DailyFeelsLike
+import com.sun.weatherapp.data.model.DailyTemp
+import com.sun.weatherapp.data.model.DailyWeather
+import com.sun.weatherapp.data.model.HourlyWeather
+import com.sun.weatherapp.data.model.Weather
+import com.sun.weatherapp.data.model.WeatherDetailResponse
+
+object TestDummies {
+    fun stubWeatherDetail(): WeatherDetailResponse {
+        val weather = listOf(Weather(id = 800, main = "Clear", description = "clear", icon = "01d"))
+        val current = CurrentWeather(
+            dt = 0L,
+            sunrise = 0L,
+            sunset = 0L,
+            temp = 300.0,
+            feels_like = 300.0,
+            pressure = 1013,
+            humidity = 50,
+            dew_point = 10.0,
+            uvi = 5.0,
+            clouds = 0,
+            visibility = 10000,
+            wind_speed = 1.0,
+            wind_deg = 0,
+            wind_gust = null,
+            weather = weather
+        )
+        val hourly = listOf(
+            HourlyWeather(
+                dt = 0L,
+                temp = 300.0,
+                feels_like = 300.0,
+                pressure = 1013,
+                humidity = 50,
+                dew_point = 10.0,
+                uvi = 5.0,
+                clouds = 0,
+                visibility = 10000,
+                wind_speed = 1.0,
+                wind_deg = 0,
+                wind_gust = null,
+                weather = weather,
+                pop = 0.0,
+                rain = null
+            )
+        )
+        val daily = listOf(
+            DailyWeather(
+                dt = 0L,
+                sunrise = 0L,
+                sunset = 0L,
+                moonrise = null,
+                moonset = null,
+                moon_phase = 0.0,
+                summary = "",
+                temp = DailyTemp(300.0, 299.0, 301.0, 298.0, 300.0, 299.0),
+                feels_like = DailyFeelsLike(300.0, 298.0, 299.0, 297.0),
+                pressure = 1013,
+                humidity = 50,
+                dew_point = 10.0,
+                wind_speed = 1.0,
+                wind_deg = 0,
+                wind_gust = null,
+                weather = weather,
+                clouds = 0,
+                pop = 0.0,
+                rain = null,
+                uvi = 5.0
+            )
+        )
+        return WeatherDetailResponse(
+            lat = 0.0,
+            lon = 0.0,
+            timezone = "UTC",
+            timezone_offset = 0,
+            current = current,
+            hourly = hourly,
+            daily = daily
+        )
+    }
+}
